+++ server/interfaces/GetCourseParams.ts
@@ -1,0 +1,5 @@
+export interface GetCourseParams {
+    termId: string
+    courseId: string
+  }
+
